Corrected installation and execution for local installations of yosys.  Specifically, yosys calls yosys-abc, which if not in the standard search path, must be specified using the "-exe" switch to the "abc" command in yosys.

+++ scripts/router.sh
@@ -153,8 +153,10 @@ if (${scripting} == "T") then
    endif
 endif
 
-rm ${rootname}.def
-mv ${rootname}_route.def ${rootname}.def
+if ( -f ${rootname}_route.def ) then
+   rm -f ${rootname}.def
+   mv ${rootname}_route.def ${rootname}.def
+endif
 
 #------------------------------------------------------------
 # Done!
